python/requirements.txt
xgboost>=2.0.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.5
tensorflow>=2.13.0
keras>=2.13.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

